feito²

+++ Pizzaria_do_Chiquinho/src/principal/Pedido.java
@@ -23,7 +23,7 @@ public class Pedido {
 	public float getTotal() {
 		return total;
 	}
-	public void setTotal(int valor) {
+	public void setTotal(float valor) {
 		this.total = valor;
 	}
 	public String getStatus() {
